+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 - Change the default shortcut for Format > Font > “Reset to Default” command to `⌘0` and let “Bigger” command accept also `⌘=`.
 - Add shortcut `⌘/` to “Show/Hide Status Bar” command.
+- Add shortcut `⌘⇧T` to “Show/Hide Tab Bar” command.
 - Avoid showing rainbow cursor when cancelling the initial syntax highlight.
 - [trivial] Modifty the layout of the progress dialog and the regular expression syntax reference.